Changed the style of some buttons in teacher component
