+++ src/app/authentication/login/login.component.html
@@ -16,4 +16,4 @@
     </div>
     <div class="col-md-2">
     </div>
-</div>>
+</div>
